+++ src/cache.h
@@ -119,7 +119,7 @@ private:
                      QString resType);
     void verifyResources(int &resourcesDeleted);
     bool removeMediaFile(Resource &res, const char *msg);
-    bool fillType(QString &type, QList<Resource> &matchingResources,
+    bool fillType(const QString &type, QList<Resource> &matchingResources,
                   QString &result, QString &source);
     bool doVideoConvert(Resource &resource, QString &cacheFile,
                         const QString &cacheAbsolutePath,