+++ test/settings/settings_test.cpp
@@ -373,8 +373,6 @@ void TestSettings::configIniFrontend() {
     QCOMPARE(config.includePattern, exp);
     exp = settings.value("maxLength");
     QCOMPARE(config.maxLength, exp);
-    exp = settings.value("manuals");
-    QCOMPARE(config.manuals, exp);
     exp = settings.value("mediaFolder");
     QCOMPARE(config.mediaFolder, exp);
     exp = settings.value("mediaFolderHidden");
@@ -464,8 +462,6 @@ void TestSettings::configIniScraper() {
     QCOMPARE(config.interactive, exp);
     exp = settings.value("jpgQuality");
     QCOMPARE(config.jpgQuality, exp);
-    exp = settings.value("manuals");
-    QCOMPARE(config.manuals, exp);
     exp = settings.value("maxLength");
     QCOMPARE(config.maxLength, exp);
     QCOMPARE(config.minMatch, 65);